Ques19 to ques25

+++ ques19.c
@@ -23,7 +23,7 @@ char grade(float percentage){
 
 float percentage(int m[]){
     int sum = m[1] + m[2] + m[3] + m[4] + m[0];
-    float per = sum / 5;
+    float per = (float)sum / 5;
     return per;
 }
